fix: profile modal username alignment
