+++ src/styles/Home.module.sass
@@ -15,10 +15,12 @@
 
 
 //Modal profile component
+.divUser
+    @apply flex items-center
 .profileImage
     @apply rounded-full border-8 border-solid border-primarycolor w-[210px] h-[210px]
 .usernameProfile
-    @apply text-headline3 font-medium text-text right-0 whitespace-nowrap overflow-hidden text-ellipsis w-[80px] text-center
+    @apply text-headline3 font-medium text-text whitespace-nowrap overflow-hidden text-ellipsis w-[99px] text-left
 .iconGame
     @apply text-text w-[50px] h-[50px] mr-2
 .gameName
